+++ juriscraper/scotus/__init__.py
@@ -1,5 +1,4 @@
 from .scotus_docket_report import SCOTUSDocketReport
+from .scotus_docket_report_html import SCOTUSDocketReportHTML
 
-__all__ = [
-    SCOTUSDocketReport,
-]
+__all__ = [SCOTUSDocketReport, SCOTUSDocketReportHTML]