+++ tests/local/test_ScotusParseDocketTest.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from juriscraper.scotus import SCOTUSDocketReport
+from juriscraper.scotus import SCOTUSDocketReport, SCOTUSDocketReportHTML
 from tests import TESTS_ROOT_EXAMPLES_SCOTUS
 from tests.local.PacerParseTestCase import PacerParseTestCase
 
@@ -15,6 +15,10 @@ class ScotusParseDocketTest(PacerParseTestCase):
     def test_parsing_scotus_json_dockets(self):
         path_root = os.path.join(TESTS_ROOT_EXAMPLES_SCOTUS, "dockets", "json")
         self.parse_files(path_root, "*.compare.json", SCOTUSDocketReport)
+
+    def test_parsing_scotus_html_dockets(self):
+        path_root = os.path.join(TESTS_ROOT_EXAMPLES_SCOTUS, "dockets", "html")
+        self.parse_files(path_root, "*.html", SCOTUSDocketReportHTML)
 
 
 class LowerCourtCasesCleaningTest(unittest.TestCase):