+++ CHANGES.md
@@ -15,7 +15,7 @@ Releases are also tagged in git, if that's helpful.
 The following changes are not yet released, but are code complete:
 
 Features:
--
+- Add scotus_docket_report_html to parse SCOTUS HTML dockets
 
 Changes:
 -